feat: add mock api to tests, clear consoles and comments
