+++ src/services/BuildClient.ts
@@ -1,11 +1,11 @@
 import fetch from 'node-fetch';
 import {
   ClientBuilder,
-  type AuthMiddlewareOptions, // Required for auth
+  type AuthMiddlewareOptions,
   type HttpMiddlewareOptions,
   ExistingTokenMiddlewareOptions,
   PasswordAuthMiddlewareOptions,
-  UserAuthOptions, // Required for sending HTTP requests
+  UserAuthOptions,
 } from '@commercetools/sdk-client-v2';
 import { createApiBuilderFromCtpClient } from '@commercetools/platform-sdk';
 import { getExistingToken, tokenCache } from '../util';
@@ -45,12 +45,10 @@ const formAnonFlow = () => {
     ? new ClientBuilder()
         .withExistingTokenFlow(currentToken, existingAuthMiddlewareOptions)
         .withHttpMiddleware(httpMiddlewareOptions)
-        .withLoggerMiddleware()
         .build()
     : new ClientBuilder()
         .withAnonymousSessionFlow(authAnonMiddlewareOptions)
         .withHttpMiddleware(httpMiddlewareOptions)
-        .withLoggerMiddleware()
         .build();
 
   return createApiBuilderFromCtpClient(ctpClient).withProjectKey({
@@ -59,7 +57,6 @@ const formAnonFlow = () => {
 };
 
 const formPassFlow = (user: UserAuthOptions) => {
-  console.log('new formPassFlow');
   localStorage.clear();
   const passwordAuthMiddlewareOptions: PasswordAuthMiddlewareOptions = {
     host: process.env.REACT_APP_AUTH_URL!,
@@ -75,10 +72,8 @@ const formPassFlow = (user: UserAuthOptions) => {
   };
 
   const newCtpClient = new ClientBuilder()
-    // .withClientCredentialsFlow(authAnonMiddlewareOptions)
     .withPasswordFlow(passwordAuthMiddlewareOptions)
     .withHttpMiddleware(httpMiddlewareOptions)
-    .withLoggerMiddleware()
     .build();
   return createApiBuilderFromCtpClient(newCtpClient).withProjectKey({
     projectKey: projectKey,
