+++ tests/RegistrationFrom.test.tsx
@@ -1,9 +1,32 @@
+import { enableFetchMocks } from 'jest-fetch-mock';
+import fetch from 'jest-fetch-mock';
+enableFetchMocks();
+
 import { fireEvent, render, waitFor, act } from '@testing-library/react';
 import { RegisterForm } from '../src/components/auth/RegisterForm';
 import userEvent from '@testing-library/user-event';
+import { AuthProvider } from '../src/context/AuthProvider';
+
+fetch.mockResponse(() =>
+  Promise.resolve({ json: () => Promise.resolve('resolved') }).then(() => {
+    return {
+      body: JSON.stringify({
+        customer: {
+          firstName: 'Test',
+        },
+      }),
+    };
+  }),
+);
+
+window.alert = () => {};
 
 test('checks Registration form render with correct inputs', async () => {
-  const wrapper = render(<RegisterForm />);
+  const wrapper = render(
+    <AuthProvider>
+      <RegisterForm />
+    </AuthProvider>,
+  );
   expect(wrapper).toBeTruthy();
 
   const inputs = {
